refactor: Money 클래스 vo 패키지로 이동
