+++ src/main/java/lotto/domain/Store.java
@@ -1,5 +1,7 @@
 package lotto.domain;
 
+import lotto.domain.vo.Money;
+
 import java.util.ArrayList;
 import java.util.List;
