+++ src/test/java/lotto/domain/MoneyTest.java
@@ -2,7 +2,8 @@ package lotto.domain;
 
 import static org.assertj.core.api.Assertions.*;
 import java.math.BigDecimal;
-import lotto.controller.Rate;
+
+import lotto.domain.vo.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
