+++ src/test/java/lotto/domain/RankTest.java
@@ -3,6 +3,8 @@ package lotto.domain;
 import static org.assertj.core.api.Assertions.*;
 import java.util.ArrayList;
 import java.util.List;
+
+import lotto.domain.vo.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
